+++ integrations/src/test/java/io/jenkins/plugins/casc/GitToolInstallationTest.java
@@ -20,7 +20,7 @@ public class GitToolInstallationTest {
     @Test
     @ConfiguredWithCode("GitToolInstallationTest.yml")
     public void configure_git_installations() throws Exception {
-        final Jenkins jenkins = Jenkins.getInstance();
+        final Jenkins jenkins = Jenkins.get();
         final GitTool.DescriptorImpl descriptor = (GitTool.DescriptorImpl) jenkins.getDescriptor(GitTool.class);
         assertEquals(2, descriptor.getInstallations().length);
         assertEquals("/usr/local/bin/git", descriptor.getInstallation("another_git").getGitExe());