+++ integrations/src/test/java/io/jenkins/plugins/casc/GithubOrganisationFolderTest.java
@@ -23,7 +23,7 @@ public class GithubOrganisationFolderTest {
     // Fails as Items do override submit() with manual data-binding implementation
     @ConfiguredWithCode("GithubOrganisationFolderTest.yml")
     public void configure_github_organisation_folder_seed_job() throws Exception {
-        final TopLevelItem job = Jenkins.getInstance().getItem("ndeloof");
+        final TopLevelItem job = Jenkins.get().getItem("ndeloof");
         assertNotNull(job);
         assertTrue(job instanceof OrganizationFolder);
         OrganizationFolder folder = (OrganizationFolder) job;