+++ integrations/src/test/java/io/jenkins/plugins/casc/LDAPSecurityRealmTest.java
@@ -2,13 +2,13 @@ package io.jenkins.plugins.casc;
 
 import hudson.security.LDAPSecurityRealm;
 import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
-import io.jenkins.plugins.casc.misc.EnvVarsRule;
 import io.jenkins.plugins.casc.misc.JenkinsConfiguredWithCodeRule;
 import jenkins.model.IdStrategy;
 import jenkins.model.Jenkins;
 import jenkins.security.plugins.ldap.LDAPConfiguration;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.contrib.java.lang.system.EnvironmentVariables;
 import org.junit.rules.RuleChain;
 
 import static org.junit.Assert.assertEquals;
@@ -20,14 +20,14 @@ import static org.junit.Assert.assertTrue;
 public class LDAPSecurityRealmTest {
 
     @Rule
-    public RuleChain chain = RuleChain.outerRule(new EnvVarsRule()
-            .env("LDAP_PASSWORD", "SECRET"))
+    public RuleChain chain = RuleChain.outerRule(new EnvironmentVariables()
+            .set("LDAP_PASSWORD", "SECRET"))
             .around(new JenkinsConfiguredWithCodeRule());
 
     @Test
     @ConfiguredWithCode("LDAPSecurityRealmTest.yml")
     public void configure_securityRealm() {
-        final Jenkins jenkins = Jenkins.getInstance();
+        final Jenkins jenkins = Jenkins.get();
         final LDAPSecurityRealm securityRealm = (LDAPSecurityRealm) jenkins.getSecurityRealm();
         assertEquals(1, securityRealm.getConfigurations().size());
         assertTrue(securityRealm.getUserIdStrategy() instanceof IdStrategy.CaseInsensitive);