+++ integrations/src/test/java/io/jenkins/plugins/casc/ProjectMatrixAuthorizationTest.java
@@ -23,8 +23,8 @@ public class ProjectMatrixAuthorizationTest {
     @Test
     @ConfiguredWithCode("ProjectMatrixStrategy.yml")
     public void checkCorrectlyConfiguredPermissions() throws Exception {
-        assertEquals("The configured instance must use the Global Matrix Authentication Strategy", ProjectMatrixAuthorizationStrategy.class, Jenkins.getInstance().getAuthorizationStrategy().getClass());
-        ProjectMatrixAuthorizationStrategy gms = (ProjectMatrixAuthorizationStrategy) Jenkins.getInstance().getAuthorizationStrategy();
+        assertEquals("The configured instance must use the Global Matrix Authentication Strategy", ProjectMatrixAuthorizationStrategy.class, Jenkins.get().getAuthorizationStrategy().getClass());
+        ProjectMatrixAuthorizationStrategy gms = (ProjectMatrixAuthorizationStrategy) Jenkins.get().getAuthorizationStrategy();
 
         List<String> adminPermission = new ArrayList<>(gms.getGrantedPermissions().get(Jenkins.ADMINISTER));
         assertEquals("authenticated", adminPermission.get(0));