+++ integrations/src/test/java/io/jenkins/plugins/casc/SystemCredentialsTest.java
@@ -8,13 +8,12 @@ import com.cloudbees.plugins.credentials.common.UsernamePasswordCredentials;
 import hudson.security.ACL;
 import io.jenkins.plugins.casc.impl.configurators.DataBoundConfigurator;
 import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
-import io.jenkins.plugins.casc.misc.EnvVarsRule;
 import io.jenkins.plugins.casc.misc.JenkinsConfiguredWithCodeRule;
 import io.jenkins.plugins.casc.model.CNode;
 import jenkins.model.Jenkins;
-import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.contrib.java.lang.system.EnvironmentVariables;
 import org.junit.rules.RuleChain;
 import org.jvnet.hudson.test.LoggerRule;
 
@@ -24,9 +23,7 @@ import java.util.logging.Level;
 import java.util.logging.LogRecord;
 
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
 
 /**
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
@@ -39,10 +36,10 @@ public class SystemCredentialsTest {
         .capture(100);
 
     @Rule
-    public RuleChain chain = RuleChain.outerRule(new EnvVarsRule()
-            .env("SUDO_PASSWORD", "1234")
-            .env("SSH_PRIVATE_KEY", "s3cr3t")
-            .env("SSH_KEY_PASSWORD", "ABCD"))
+    public RuleChain chain = RuleChain.outerRule(new EnvironmentVariables()
+            .set("SUDO_PASSWORD", "1234")
+            .set("SSH_PRIVATE_KEY", "s3cr3t")
+            .set("SSH_KEY_PASSWORD", "ABCD"))
             .around(log)
             .around(new JenkinsConfiguredWithCodeRule());
 
@@ -50,7 +47,7 @@ public class SystemCredentialsTest {
     @Test
     @ConfiguredWithCode("SystemCredentialsTest.yml")
     public void configure_system_credentials() throws Exception {
-        Jenkins jenkins = Jenkins.getInstance();
+        Jenkins jenkins = Jenkins.get();
 
         List<UsernamePasswordCredentials> ups = CredentialsProvider.lookupCredentials(
                 UsernamePasswordCredentials.class, jenkins, ACL.SYSTEM, Collections.emptyList()