+++ plugin/src/main/java/io/jenkins/plugins/casc/core/AdminWhitelistRuleConfigurator.java
@@ -27,6 +27,7 @@ import io.jenkins.plugins.casc.BaseConfigurator;
 import io.jenkins.plugins.casc.ConfigurationContext;
 import io.jenkins.plugins.casc.model.CNode;
 import io.jenkins.plugins.casc.model.Mapping;
+import java.util.Collections;
 import jenkins.model.Jenkins;
 import jenkins.security.s2m.AdminWhitelistRule;
 import org.kohsuke.accmod.Restricted;
@@ -64,16 +65,18 @@ public class AdminWhitelistRuleConfigurator extends BaseConfigurator<AdminWhitel
 
     @Override
     public Set<Attribute<AdminWhitelistRule,?>> describe() {
-        return new HashSet<>(Arrays.asList(
-                new Attribute<AdminWhitelistRule, Boolean>("enabled", Boolean.class)
-                        .getter(target -> !target.getMasterKillSwitch())
-                        .setter((target, value) -> target.setMasterKillSwitch(!(Boolean)value))
+        return new HashSet<>(Collections.singletonList(
+            new Attribute<AdminWhitelistRule, Boolean>("enabled", Boolean.class)
+                .getter(target -> !target.getMasterKillSwitch())
+                .setter((target, value) -> target.setMasterKillSwitch(!value))
         ));
     }
 
     @CheckForNull
     @Override
     public CNode describe(AdminWhitelistRule instance, ConfigurationContext context) {
-        return null; // FIXME
+        final Mapping mapping = new Mapping();
+        mapping.put("enabled", !instance.getMasterKillSwitch());
+        return mapping;
     }
 }