+++ plugin/src/main/java/io/jenkins/plugins/casc/core/JNLPLauncherConfigurator.java
@@ -26,7 +26,7 @@ public class JNLPLauncherConfigurator extends DataBoundConfigurator<JNLPLauncher
         try {
             return super.instance(config, context);
         } catch (ConfiguratorException e) {
-            // see https://issues.jenkins-ci.org/browse/JENKINS-51603
+            // see https://issues.jenkins.io/browse/JENKINS-51603
             final CNode tunnel = config.get("tunnel");
             final CNode vmargs = config.get("vmargs");
             return new JNLPLauncher(tunnel != null ? tunnel.asScalar().getValue() : null,