+++ plugin/src/main/java/io/jenkins/plugins/casc/impl/secrets/VaultSecretSource.java
@@ -5,8 +5,11 @@ import com.bettercloud.vault.VaultConfig;
 import com.bettercloud.vault.VaultException;
 import hudson.Extension;
 import io.jenkins.plugins.casc.SecretSource;
-import org.kohsuke.accmod.Restricted;
-import org.kohsuke.accmod.restrictions.Beta;
+import io.jenkins.plugins.casc.impl.secrets.vault.VaultAppRoleAuthenticator;
+import io.jenkins.plugins.casc.impl.secrets.vault.VaultAuthenticator;
+import io.jenkins.plugins.casc.impl.secrets.vault.VaultSingleTokenAuthenticator;
+import io.jenkins.plugins.casc.impl.secrets.vault.VaultUserPassAuthenticator;
+import org.apache.commons.lang.StringUtils;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -18,72 +21,153 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
- * Requires either CASC_VAULT_USER and CASC_VAULT_PW, or CASC_VAULT_TOKEN environment variables set
- * alongside with CASC_VAULT_PATH and CASC_VAULT_URL
+ * Requires either CASC_VAULT_USER and CASC_VAULT_PW, or CASC_VAULT_TOKEN,
+ * or CASC_VAULT_APPROLE and CASC_VAULT_APPROLE_SECRET
+ * environment variables set alongside with CASC_VAULT_PATHS and CASC_VAULT_URL
  */
 @Extension
 
 public class VaultSecretSource extends SecretSource {
 
     private final static Logger LOGGER = Logger.getLogger(VaultSecretSource.class.getName());
+
+    private static final String CASC_VAULT_FILE = "CASC_VAULT_FILE";
+    private static final String CASC_VAULT_PW = "CASC_VAULT_PW";
+    private static final String CASC_VAULT_USER = "CASC_VAULT_USER";
+    private static final String CASC_VAULT_URL = "CASC_VAULT_URL";
+    private static final String CASC_VAULT_MOUNT = "CASC_VAULT_MOUNT";
+    private static final String CASC_VAULT_TOKEN = "CASC_VAULT_TOKEN";
+    private static final String CASC_VAULT_APPROLE = "CASC_VAULT_APPROLE";
+    private static final String CASC_VAULT_APPROLE_SECRET = "CASC_VAULT_APPROLE_SECRET";
+    private static final String CASC_VAULT_NAMESPACE = "CASC_VAULT_NAMESPACE";
+    private static final String CASC_VAULT_ENGINE_VERSION = "CASC_VAULT_ENGINE_VERSION";
+    private static final String CASC_VAULT_PATHS = "CASC_VAULT_PATHS";
+    private static final String CASC_VAULT_PATH = "CASC_VAULT_PATH"; // TODO: deprecate!
+    private static final String DEFAULT_ENGINE_VERSION = "2";
+    private static final String DEFAULT_USER_BACKEND = "userpass";
+
     private Map<String, String> secrets = new HashMap<>();
+    private Vault vault;
+    private VaultConfig vaultConfig;
+    private VaultAuthenticator vaultAuthenticator;
+    private String[] vaultPaths;
 
     public VaultSecretSource() {
-        String vaultFile = System.getenv("CASC_VAULT_FILE");
+        Optional<String> vaultFile = Optional.ofNullable(System.getenv(CASC_VAULT_FILE));
         Properties prop = new Properties();
-        if (vaultFile != null) {
-            try (FileInputStream input = new FileInputStream(vaultFile)) {
-                prop.load(input);
-                if (prop.isEmpty()) {
-                    LOGGER.log(Level.WARNING, "Vault secret file is empty");
-                }
-            } catch (IOException ex) {
-                LOGGER.log(Level.WARNING, "Failed to load Vault secrets from file", ex);
-            }
-        }
-        String vaultPw = getVariable("CASC_VAULT_PW", prop);
-        String vaultUsr = getVariable("CASC_VAULT_USER", prop);
-        String vaultPth = getVariable("CASC_VAULT_PATH", prop);
-        String vaultUrl = getVariable("CASC_VAULT_URL", prop);
-        String vaultMount = getVariable("CASC_VAULT_MOUNT", prop);
-        String vaultToken = getVariable("CASC_VAULT_TOKEN", prop);
-        String vaultAppRole = getVariable("CASC_VAULT_APPROLE", prop);
-        String vaultAppRoleSecret = getVariable("CASC_VAULT_APPROLE_SECRET", prop);
+        vaultFile.ifPresent(file -> readPropertiesFromVaultFile(file, prop));
 
-        if(((vaultPw != null && vaultUsr != null) || 
-            vaultToken != null || 
-            (vaultAppRole != null && vaultAppRoleSecret != null)) && vaultPth != null && vaultUrl != null) {
+        // Parse variables
+        Optional<String> vaultEngineVersionOpt = getVariable(CASC_VAULT_ENGINE_VERSION, prop);
+        Optional<String> vaultUrl = getVariable(CASC_VAULT_URL, prop);
+        Optional<String> vaultNamespace = getVariable(CASC_VAULT_NAMESPACE, prop);
+        Optional<String[]> vaultPaths = getCommaSeparatedVariables(CASC_VAULT_PATHS, prop)
+                .map(Optional::of)
+                .orElseGet(() -> getCommaSeparatedVariables(CASC_VAULT_PATH, prop)); // TODO: deprecate!
+
+        // Check mandatory variables are set
+        if (!vaultUrl.isPresent() || !vaultPaths.isPresent()) return;
+
+        String vaultEngineVersion = vaultEngineVersionOpt.orElse(DEFAULT_ENGINE_VERSION);
+        this.vaultPaths = vaultPaths.get();
+        determineAuthenticator(prop);
+
+        // configure vault client
+        vaultConfig = new VaultConfig().address(vaultUrl.get());
+        try {
             LOGGER.log(Level.FINE, "Attempting to connect to Vault: {0}", vaultUrl);
+            if (vaultNamespace.isPresent()) {
+                vaultConfig.nameSpace(vaultNamespace.get());
+                LOGGER.log(Level.FINE, "Using namespace with Vault: {0}", vaultNamespace);
+            }
+
+            vaultConfig.engineVersion(Integer.parseInt(vaultEngineVersion));
+            LOGGER.log(Level.FINE, "Using engine version: {0}", vaultEngineVersion);
+        } catch (VaultException e) {
+            LOGGER.log(Level.WARNING, "Could not configure vault connection", e);
+        }
+
+        try {
+            vaultConfig.build();
+        } catch (VaultException e) {
+            LOGGER.log(Level.WARNING, "Could not configure vault client", e);
+        }
+
+        vault = new Vault(vaultConfig);
+    }
+
+    private void determineAuthenticator(Properties prop) {
+        Optional<String> vaultPw = getVariable(CASC_VAULT_PW, prop);
+        Optional<String> vaultUser = getVariable(CASC_VAULT_USER, prop);
+        Optional<String> vaultMount = getVariable(CASC_VAULT_MOUNT, prop);
+        Optional<String> vaultToken = getVariable(CASC_VAULT_TOKEN, prop);
+        Optional<String> vaultAppRole = getVariable(CASC_VAULT_APPROLE, prop);
+        Optional<String> vaultAppRoleSecret = getVariable(CASC_VAULT_APPROLE_SECRET, prop);
+
+        vaultToken.ifPresent(
+                token -> vaultAuthenticator = new VaultSingleTokenAuthenticator(token)
+        );
+
+        vaultUser.ifPresent(
+                user -> vaultPw.ifPresent(
+                        pw -> vaultAuthenticator = new VaultUserPassAuthenticator(user, pw, vaultMount.orElse(DEFAULT_USER_BACKEND))
+                )
+        );
+
+        vaultAppRole.ifPresent(
+                approle -> vaultAppRoleSecret.ifPresent(
+                        approleSecret -> vaultAuthenticator = new VaultAppRoleAuthenticator(approle, approleSecret)
+                )
+        );
+
+        if (vaultAuthenticator == null) {
+            LOGGER.log(Level.WARNING, "Could not determine vault authentication method. Not able to read secrets from vault.");
+        }
+    }
+
+    private void readSecretsFromVault() {
+        Optional<String[]> vaultPathsOpt = Optional.ofNullable(vaultPaths);
+
+        if (vaultPathsOpt.isPresent()) {
             try {
-                VaultConfig config = new VaultConfig().address(vaultUrl).build();
-                Vault vault = new Vault(config);
-                //Obtain a login token
-                final String token;
-                if (vaultToken != null) {
-                    token = vaultToken;
-                    LOGGER.log(Level.FINE, "Using supplied token to access Vault");
-                } else if (vaultAppRole != null && vaultAppRoleSecret != null) {
-                    token = vault.auth().loginByAppRole(vaultAppRole, vaultAppRoleSecret).getAuthClientToken();
-                    LOGGER.log(Level.FINE, "Login to Vault using AppRole/SecretID successful");
-                } else {
-                    token = vault.auth().loginByUserPass(vaultUsr, vaultPw, vaultMount).getAuthClientToken();
-                    LOGGER.log(Level.FINE, "Login to Vault using U/P successful");
+                // refresh map
+                secrets = new HashMap<>();
+
+                // Parse secrets
+                for (String vaultPath : vaultPathsOpt.get()) {
+
+                    // check if we overwrite an existing key from another path
+                    Map<String, String> nextSecrets = vault.logical().read(vaultPath).getData();
+                    for (String key : nextSecrets.keySet()) {
+                        if (secrets.containsKey(key)) {
+                            LOGGER.log(Level.WARNING, "Key {0} exists in multiple vault paths.", key);
+                        }
+                    }
+
+                    // merge
+                    secrets.putAll(nextSecrets);
                 }
-                config.token(token).build();
-                secrets = vault.logical().read(vaultPth).getData();
-            } catch (VaultException ve) {
-                LOGGER.log(Level.WARNING, "Unable to connect to Vault", ve);
+            } catch (VaultException e) {
+                LOGGER.log(Level.WARNING, "Unable to fetch secret from Vault", e);
             }
         }
     }
 
-    @Override
-    public Optional<String> reveal(String vaultKey) {
-        Optional<String> returnValue = Optional.empty();
-        if(secrets.containsKey(vaultKey))  {
-            returnValue = Optional.of(secrets.get(vaultKey));
+    private void readPropertiesFromVaultFile(String vaultFile, Properties prop) {
+        try (FileInputStream input = new FileInputStream(vaultFile)) {
+            prop.load(input);
+            if (prop.isEmpty()) {
+                LOGGER.log(Level.WARNING, "Vault secret file is empty");
+            }
+        } catch (IOException ex) {
+            LOGGER.log(Level.WARNING, "Failed to load Vault secrets from file", ex);
         }
-        return returnValue;
+    }
+
+    @Override
+    public Optional<String> reveal(String secret) {
+        if (StringUtils.isBlank(secret)) return Optional.empty();
+        return Optional.ofNullable(secrets.get(secret));
     }
 
     public Map<String, String> getSecrets() {
@@ -94,11 +178,27 @@ public class VaultSecretSource extends SecretSource {
         this.secrets = secrets;
     }
 
-    private String getVariable(String key, Properties prop) {
-        if (prop != null && !prop.isEmpty()) {
-            return prop.getProperty(key, System.getenv(key));
-        } else {
-            return System.getenv(key);
+    private Optional<String> getVariable(String key, Properties prop) {
+        return Optional.ofNullable(prop.getProperty(key, System.getenv(key)));
+    }
+
+    private Optional<String[]> getCommaSeparatedVariables(String key, Properties prop) {
+        if (key.equals(CASC_VAULT_PATH))
+            LOGGER.log(Level.WARNING, "[Deprecation Warning] CASC_VAULT_PATH will be deprecated. " +
+                    "Please use CASC_VAULT_PATHS instead."); // TODO: deprecate!
+        return getVariable(key, prop).map(str -> str.split(","));
+    }
+
+    @Override
+    public void init() {
+        // Ensure secrets are up-to-date and Check vault authentication
+        if (vaultAuthenticator != null) {
+            try {
+                vaultAuthenticator.authenticate(vault, vaultConfig);
+            } catch (VaultException e) {
+                LOGGER.log(Level.WARNING, "Could not authenticate with vault client", e);
+            }
+            readSecretsFromVault();
         }
     }
 }