+++ plugin/src/test/java/io/jenkins/plugins/casc/misc/EnvsFromFile.java
@@ -1,0 +1,19 @@
+package io.jenkins.plugins.casc.misc;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+import org.junit.Test;
+
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target(METHOD)
+@Retention(RUNTIME)
+public @interface EnvsFromFile {
+
+    String[] value() default {};
+
+    Class<? extends Throwable> expected() default Test.None.class;
+
+    String message() default "";
+}